add utils file to organize functions, add persistance, updated lsp config to handle learner-frontend project well
